[utility] internalize utility file
