+++ include/fcarouge/kalman.hpp
@@ -50,7 +50,7 @@ For more information, please refer to <https://unlicense.org> */
 #include "kalman_internal/factory.hpp"
 #include "kalman_internal/format.hpp"
 #include "kalman_internal/printer.hpp"
-#include "utility.hpp"
+#include "kalman_internal/utility.hpp"
 
 namespace fcarouge {
 //! @name Types
