+++ include/fcarouge/kalman_internal/x_z_p_q_r_hh_us_ps.hpp
@@ -39,8 +39,8 @@ For more information, please refer to <https://unlicense.org> */
 #ifndef FCAROUGE_KALMAN_INTERNAL_X_Z_P_Q_R_HH_US_PS_HPP
 #define FCAROUGE_KALMAN_INTERNAL_X_Z_P_Q_R_HH_US_PS_HPP
 
-#include "fcarouge/utility.hpp"
 #include "function.hpp"
+#include "utility.hpp"
 
 #include <tuple>
