+++ include/fcarouge/kalman_internal/x_z_p_qq_rr_f.hpp
@@ -39,8 +39,8 @@ For more information, please refer to <https://unlicense.org> */
 #ifndef FCAROUGE_KALMAN_INTERNAL_X_Z_P_QQ_RR_F_HPP
 #define FCAROUGE_KALMAN_INTERNAL_X_Z_P_QQ_RR_F_HPP
 
-#include "fcarouge/utility.hpp"
 #include "function.hpp"
+#include "utility.hpp"
 
 #include <tuple>
