+++ include/fcarouge/kalman_internal/x_z_p_r.hpp
@@ -39,7 +39,7 @@ For more information, please refer to <https://unlicense.org> */
 #ifndef FCAROUGE_KALMAN_INTERNAL_X_Z_P_R_HPP
 #define FCAROUGE_KALMAN_INTERNAL_X_Z_P_R_HPP
 
-#include "fcarouge/utility.hpp"
+#include "utility.hpp"
 
 namespace fcarouge::kalman_internal {
 template <typename Type> struct x_z_p_r {
